# docs/api-contract.yaml
openapi: 3.0.0
info:
  title: Tactical RAG API
  version: 1.0.0
  description: |
    REST API for the Tactical RAG system - an enterprise-grade Retrieval Augmented Generation system
    with adaptive retrieval strategies and Redis caching.
  contact:
    name: API Support
    email: support@tacticalrag.com
servers:
  - url: http://localhost:8000
    description: Development server
  - url: http://localhost:8000/api
    description: Development API endpoint
...
paths:
...
  /api/query:
    post:
      summary: Submit a query
      description: Process a question using the RAG system with optional conversation context
      tags:
        - Query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
                - mode
              properties:
                question:
                  type: string
                  description: The question to ask
                  example: "What are the main requirements for beard grooming in AR 670-1?"
                mode:
                  type: string
                  enum: [simple, adaptive]
                  description: Retrieval mode to use
                  example: adaptive
                use_context:
                  type: boolean
                  default: true
                  description: Whether to use conversation context
                session_id:
                  type: string
                  description: Optional session ID for conversation tracking
      responses:
        '200':
          description: Query result
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                    description: Generated answer
                  sources:
                    type: array
                    items:
                      $ref: '#/components/schemas/Source'
                  metadata:
                    type: object
                    properties:
                      strategy_used:
                        type: string
                        enum: [simple_dense, hybrid_search, advanced_multi_query]
                      query_type:
                        type: string
                        enum: [simple, moderate, complex]
                      mode:
                        type: string
                        enum: [simple, adaptive]
                      latency_ms:
                        type: number
                        format: float
                      cache_hit:
                        type: boolean
                  explanation:
                    type: object
                    nullable: true
                    properties:
                      query_type:
                        type: string
                      complexity_score:
                        type: number
                      scoring_breakdown:
                        type: object
                      strategy_reasoning:
                        type: string
                  error:
                    type: boolean
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/conversation/clear:
    post:
      summary: Clear conversation history
      description: Clear conversation memory for the current session
      tags:
        - Query
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Conversation cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
components:
  schemas:
    Source:
      type: object
      properties:
        file_name:
          type: string
        file_type:
          type: string
        relevance_score:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
        excerpt:
          type: string
          description: Text excerpt from the source
        metadata:
          type: object
          properties:
            page_number:
              type: integer
              nullable: true
            chunk_index:
              type: integer
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        details:
          type: object
          nullable: true
          description: Additional error details
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for authentication (future use)
security: []